requirements.txt
apache-airflow==3.0.3
clickhouse_driver==0.2.9
numpy==1.24.3
numpy==1.23.5
pandas==1.5.3
python-dotenv==1.1.1
pytz==2022.7
Requests==2.32.4
